Use uint16_t for our various header sizes in bytes
PacketBuffer uses uint16_t for all of its sizes already.  It's easier
to tell that we're not overflowing the buffer if we know our header
sizes also fit in uint16_t.

+++ src/transport/BLE.cpp
@@ -138,7 +138,7 @@ CHIP_ERROR BLE::SendMessage(const PacketHeader & header, Header::Flags payloadFl
 {
     CHIP_ERROR err          = CHIP_NO_ERROR;
     const size_t headerSize = header.EncodeSizeBytes();
-    size_t actualEncodedHeaderSize;
+    uint16_t actualEncodedHeaderSize;
 
     VerifyOrExit(address.GetTransportType() == Type::kBle, err = CHIP_ERROR_INVALID_ARGUMENT);
     VerifyOrExit(mState == State::kInitialized, err = CHIP_ERROR_INCORRECT_STATE);
